Books response codes

+++ app/Http/Controllers/BookController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Book;
+use Illuminate\Support\Facades\Validator;
+
 use App\Models\Author;
 
 class BookController extends Controller
@@ -14,11 +16,29 @@ class BookController extends Controller
 
     public function getBook($id) {
         $book = Book::find($id);
-
-        return response()->json($book, 200);
+        if($book) {
+            return response()->json($book, 200);
+        } else {
+            return response()->json(['message' => 'not found'], 404);
+        }
     }
 
     public function addBook(Request $request) {
+
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'title' => 'required|string',
+                'isbn' => 'required|string',
+                'year'   => 'required|string',
+                'author_id' => 'required|numeric'
+            ]
+        );
+
+        if ($validator->fails()) {
+            return response()->json(["validation_errors" => $validator->errors()], 400);
+        }
+
         $book = Book::create([
             'title'     => $request->title,
             'isbn'      => $request->isbn,
@@ -30,21 +50,44 @@ class BookController extends Controller
     }
 
     public function editBook(Request $request) {
-        $book = Book::find($request->id);
-        $book->update([
-            'title'     => $request->title,
-            'year'      => $request->year,
-            'isbn'      => $request->isbn,
-            'author_id' => $request->author_id
-        ]);
+        $validator = Validator::make(
+            $request->all(),
+            [
+                'title' => 'required|string',
+                'isbn' => 'required|string',
+                'year'   => 'required|string',
+                'author_id' => 'required|numeric'
+            ]
+        );
 
-        return response()->json($book, 200);
+        if ($validator->fails()) {
+            return response()->json(["validation_errors" => $validator->errors()], 400);
+        }
+
+        $book = Book::find($request->id);
+
+        if($book) {
+            $book->update([
+                'title'     => $request->title,
+                'year'      => $request->year,
+                'isbn'      => $request->isbn,
+                'author_id' => $request->author_id
+            ]);
+            return response()->json($book, 200);
+        } else {
+            return response()->json(['message' => 'not found'], 404);
+        }
+
     }
 
     public function deleteBook($id) {
         $book = Book::find($id);
-        $book->delete();
 
-        return response()->json(['message' => 'book deleted'], 200);
+        if($book) {
+            $book->delete();
+            return response()->json(['message' => 'book deleted'], 200);
+        } else {
+            return response()->json(['message' => 'not found'], 404);
+        }
     }
 }
